requirements.txt
alembic==1.9.2
Babel==2.11.0
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
cssselect2==0.7.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
flask-weasyprint==1.0.0
Flask-WTF==1.1.0
fonttools==4.38.0
greenlet==2.0.1
html5lib==1.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
Pillow==9.4.0
pycparser==2.21
pydyf==0.5.0
PyMySQL==1.0.2
pyphen==0.13.2
python-dotenv==0.21.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.46
tinycss2==1.2.1
weasyprint==57.2
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==2.3.3
WTForms-SQLAlchemy==0.3
zipp==3.11.0
zopfli==0.2.2
